adding icons and debugging db
added houses and debugged columns in sqlite
