+++ app/src/main/AndroidManifest.xml
@@ -64,7 +64,6 @@
         <activity android:name=".TransporterMessages" />
         <activity android:name=".TransporterEditProfile" />
         <activity android:name=".TransporterHistory" />
-        <activity android:name=".TransporterSetAvailabilityEnterInfo" />
         <activity android:name=".TransporterSetAvailabilityPickDate" />
         <activity android:name=".TransporterSetAvailabilityLocationType" />
         <activity android:name=".FarmerMessages" />
