+++ app/src/main/res/layout/activity_transporter_set_availability_location_type.xml
@@ -20,13 +20,13 @@
         tools:layout_editor_absoluteX="0dp" />
 
     <Button
-        android:id="@+id/transporterSetAvailabilityLocationTypeFinishButton"
+        android:id="@+id/transporterSetAvailabilityLocationTypeNextButton"
         style="@style/Widget.AppCompat.Button.Colored"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginBottom="8dp"
         android:layout_marginEnd="8dp"
-        android:text="Finish"
+        android:text="Next"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent" />
