+++ app/src/main/res/layout/activity_transporter_set_availability_pick_date.xml
@@ -35,13 +35,13 @@
         app:layout_constraintStart_toStartOf="parent" />
 
     <Button
-        android:id="@+id/transporterSetAvailabilityDoneButton"
+        android:id="@+id/transporterSetAvailabilityFinishButton"
         style="@style/Widget.AppCompat.Button.Colored"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginBottom="8dp"
         android:layout_marginEnd="136dp"
-        android:text="Done"
+        android:text="Finish"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent" />
